# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TaskHo API",
    "version": "0.1.0",
    "description": "TaskHo backend API documentation. Authenticated endpoints require JWT Bearer token obtained from /api/auth/login or /api/auth/register."
  },
  "servers": [{ "url": "http://localhost:4000/api" }, { "url": "https://task-ho-backend.vercel.app/api" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "tab": { "type": "string", "enum": ["inbox", "today", "week", "month", "later", "waiting", "done"] },
          "status": { "type": "string", "enum": ["idle", "inprogress", "later", "waiting", "done"] },
          "customer": { "type": "string" },
          "title": { "type": "string" },
          "type": { "type": "string", "enum": ["Fatura", "Rapor", "Ödeme"] },
          "due": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "priority": { "type": "string", "enum": ["High", "Medium", "Low"] },
          "notes": { "type": "string" }
        }
      },
      "Fee": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "customer": { "type": "string" },
          "month": { "type": "string", "pattern": "^\\d{4}-\\d{2}$" },
          "amount": { "type": "number" },
          "status": { "type": "string", "enum": ["Açık", "Ödendi"] },
          "note": { "type": "string" }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "isPaid": { "type": "boolean" },
          "fee": { "type": "number" }
        }
      },
      "CalendarNote": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "customer": { "type": "string" },
          "title": { "type": "string" },
          "date": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "notes": { "type": "string" },
          "isCompleted": { "type": "boolean" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["email", "password", "name"], "properties": { "email": { "type": "string" }, "password": { "type": "string" }, "name": { "type": "string" } } }
            }
          }
        },
        "responses": { "200": { "description": "User created and token returned" } }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["email", "password"], "properties": { "email": { "type": "string" }, "password": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Returns user + token" }, "401": { "description": "Invalid credentials" } }
      }
    },
    "/auth/me": { "get": { "summary": "Current user", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" }, "401": { "description": "Unauthorized" } } } },
    "/auth/change-password": {
      "post": {
        "summary": "Change password",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["currentPassword", "newPassword"], "properties": { "currentPassword": { "type": "string" }, "newPassword": { "type": "string" } } }
            }
          }
        },
        "responses": { "200": { "description": "Changed" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/auth/request-reset": {
      "post": {
        "summary": "Request password reset token",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["email"], "properties": { "email": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Always returns message" } }
      }
    },
    "/auth/reset-password": {
      "post": {
        "summary": "Reset password with token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "resetToken", "newPassword"],
                "properties": { "email": { "type": "string" }, "resetToken": { "type": "string" }, "newPassword": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Reset" }, "400": { "description": "Invalid token" } }
      }
    },
    "/auth/profile": {
      "patch": {
        "summary": "Update profile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "email": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Updated" }, "401": { "description": "Unauthorized" } }
      }
    },

    "/tasks": {
      "get": {
        "summary": "List tasks (filters via query)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "tab", "in": "query", "schema": { "type": "string" } },
          { "name": "customer", "in": "query", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } },
          { "name": "overdue", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } }
        ],
        "responses": { "200": { "description": "Array of tasks" } }
      },
      "post": {
        "summary": "Create task",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } } },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/tasks/weekly": {
      "get": {
        "summary": "Get weekly calendar data (tasks grouped by customer and date)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "startDate", "in": "query", "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" }, "description": "Week start date (Monday), defaults to current week" }
        ],
        "responses": {
          "200": {
            "description": "Weekly calendar data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "startDate": { "type": "string", "description": "Week start (Monday)" },
                    "endDate": { "type": "string", "description": "Week end (Sunday)" },
                    "tasks": {
                      "type": "object",
                      "description": "Tasks grouped by customer name and date",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string" },
                              "title": { "type": "string" },
                              "type": { "type": "string" },
                              "status": { "type": "string" },
                              "priority": { "type": "string" },
                              "notes": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "patch": {
        "summary": "Update task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },

    "/fees": {
      "get": {
        "summary": "List fees",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "customer", "in": "query", "schema": { "type": "string" } },
          { "name": "month", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Array of fees" } }
      },
      "post": {
        "summary": "Create fee",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Fee" } } } },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/fees/{id}": {
      "patch": {
        "summary": "Update fee",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete fee",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Deleted" } }
      }
    },

    "/customers": {
      "get": {
        "summary": "List customers (search by q)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "q", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Array of customers" } }
      },
      "post": {
        "summary": "Create or update customer by name",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Customer" } } } },
        "responses": { "201": { "description": "Created" }, "200": { "description": "Updated existing" } }
      }
    },

    "/calendar": {
      "get": {
        "summary": "List calendar notes with optional date range filter",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "startDate", "in": "query", "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" } },
          { "name": "customer", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Array of calendar notes" } }
      },
      "post": {
        "summary": "Create calendar note",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CalendarNote" } } } },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/calendar/weekly": {
      "get": {
        "summary": "Get weekly calendar notes grouped by customer and date",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "startDate", "in": "query", "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" }, "description": "Start date (Monday). If not provided, current week is used." }
        ],
        "responses": {
          "200": {
            "description": "Weekly calendar data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "startDate": { "type": "string" },
                    "endDate": { "type": "string" },
                    "notes": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/CalendarNote" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/calendar/{id}": {
      "get": {
        "summary": "Get calendar note by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Calendar note" }, "404": { "description": "Not found" } }
      },
      "put": {
        "summary": "Update calendar note",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete calendar note",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    }
  }
}